Move WASM loader into module
